Refactor prediction_utils
