+++ src/careamics/prediction/__init__.py
@@ -1,0 +1,9 @@
+"""Prediction functions."""
+
+__all__ = [
+    "stitch_prediction",
+    "tta_backward",
+    "tta_forward",
+]
+
+from .prediction_utils import stitch_prediction, tta_backward, tta_forward
