+++ src/careamics/utils/__init__.py
@@ -11,10 +11,12 @@ __all__ = [
     "check_tiling_validity",
     "cwd",
     "compile_model",
+    "MetricTracker",
 ]
 
 
 from .context import cwd
+from .metrics import MetricTracker
 from .normalization import denormalize, normalize
 from .torch_utils import compile_model, get_device, setup_cudnn_reproducibility
 from .validators import check_array_validity, check_axes_validity, check_tiling_validity
